# src/openapi-spec.json
{
  "openapi": "3.0.0",
  "paths": {
    "/chat/streamInference": {
      "get": {
        "operationId": "streamInference",
        "parameters": [
          {
            "name": "prompt",
            "required": true,
            "in": "query",
            "description": "The prompt to initiate the message stream",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "conversationId",
            "required": true,
            "in": "query",
            "description": "Optional. The conversation to add the passed in prompt and llm response to.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "text/event-stream": {
                "schema": {
                  "type": "string",
                  "example": "data: { \"text\": \"response\", \"end\": \"true\" }"
                }
              }
            }
          }
        },
        "summary": "Stream a message based on a prompt",
        "tags": [
          "Chat"
        ]
      }
    },
    "/conversations/conversation/{conversationId}": {
      "get": {
        "operationId": "getConversation",
        "parameters": [
          {
            "name": "conversationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The conversation details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Conversation"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Get conversation by ID",
        "tags": [
          "Conversation"
        ]
      },
      "put": {
        "operationId": "updateConversation",
        "parameters": [
          {
            "name": "conversationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "The updated conversation object",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Conversation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The conversation has been successfully updated."
          }
        },
        "summary": "Update an existing conversation",
        "tags": [
          "Conversation"
        ]
      },
      "delete": {
        "operationId": "deleteConversation",
        "parameters": [
          {
            "name": "conversationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The conversation has been successfully deleted."
          }
        },
        "summary": "Delete a conversation",
        "tags": [
          "Conversation"
        ]
      }
    },
    "/conversations/conversation": {
      "post": {
        "operationId": "createConversation",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "The conversation to create",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConversation"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The conversation has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Conversation"
                }
              }
            }
          }
        },
        "summary": "Create a new conversation",
        "tags": [
          "Conversation"
        ]
      }
    },
    "/conversations/{conversationId}/messages/message": {
      "post": {
        "operationId": "addMessage",
        "parameters": [
          {
            "name": "conversationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMessage"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The message was successfully added",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          }
        },
        "summary": "Add a message to a conversation",
        "tags": [
          "Conversation"
        ]
      }
    },
    "/conversations/conversations/member": {
      "get": {
        "operationId": "getConversationsForMember",
        "parameters": [],
        "responses": {
          "200": {
            "description": "The conversation has been successfully deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Conversation"
                  }
                }
              }
            }
          }
        },
        "summary": "get all conversations for a member",
        "tags": [
          "Conversation"
        ]
      }
    },
    "/datasources/datasource": {
      "post": {
        "operationId": "createDatasource",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatasource"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The datasource has been created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Datasource"
                }
              }
            }
          }
        },
        "summary": "Create a new datasource",
        "tags": [
          "Datasources"
        ]
      }
    },
    "/datasources/datasource/{datasourceId}/document": {
      "post": {
        "operationId": "createDocument",
        "parameters": [
          {
            "name": "datasourceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocument"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The document has been created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          }
        },
        "summary": "Create a new document",
        "tags": [
          "Datasources"
        ]
      }
    },
    "/datasources/datasource/documents/{documentId}": {
      "get": {
        "operationId": "getDocumentById",
        "parameters": [
          {
            "name": "documentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The document details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Document"
                }
              }
            }
          },
          "404": {
            "description": "Document not found"
          }
        },
        "summary": "Get document by ID",
        "tags": [
          "Datasources"
        ]
      }
    },
    "/datasources/datasource/{datasourceId}": {
      "get": {
        "operationId": "getDatasourceById",
        "parameters": [
          {
            "name": "datasourceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The datasource details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Datasource"
                }
              }
            }
          },
          "404": {
            "description": "Datasource not found"
          }
        },
        "summary": "Get datasource by ID",
        "tags": [
          "Datasources"
        ]
      },
      "delete": {
        "operationId": "deleteDatasource",
        "parameters": [
          {
            "name": "datasourceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "indication if deleted successfully"
          }
        },
        "summary": "deletes a datasource",
        "tags": [
          "Datasources"
        ]
      }
    },
    "/datasources/datasource/{datasourceId}/documents": {
      "get": {
        "operationId": "getDocumentsForDatasource",
        "parameters": [
          {
            "name": "datasourceId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of documents for the datasource",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Document"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all documents for a datasource",
        "tags": [
          "Datasources"
        ]
      }
    },
    "/datasources/conversation/{conversationId}": {
      "get": {
        "operationId": "getDatasourcesForConversation",
        "parameters": [
          {
            "name": "conversationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of datasources for the conversation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Datasource"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all datasources for a conversation",
        "tags": [
          "Datasources"
        ]
      }
    },
    "/datasources/conversation/{conversationId}/documents": {
      "get": {
        "operationId": "getDocumentsForConversation",
        "parameters": [
          {
            "name": "conversationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of documents for the conversation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Document"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all documents for a conversation",
        "tags": [
          "Datasources"
        ]
      }
    },
    "/datasources/member": {
      "get": {
        "operationId": "getAllDatasourcesForMember",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of datasources owned by the member",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Datasource"
                  }
                }
              }
            }
          }
        },
        "summary": "get all datasources for a member",
        "tags": [
          "Datasources"
        ]
      }
    },
    "/openapi": {
      "get": {
        "operationId": "getOpenApiSpec",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "OpenApi"
        ]
      }
    }
  },
  "info": {
    "title": "AI API",
    "description": "API to manage to interact with AI",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "Message": {
        "type": "object",
        "properties": {
          "messageId": {
            "type": "string"
          },
          "sentByMemberId": {
            "type": "string"
          },
          "messageText": {
            "type": "string"
          },
          "createdDate": {
            "type": "string"
          }
        },
        "required": [
          "messageId",
          "sentByMemberId",
          "messageText",
          "createdDate"
        ]
      },
      "Conversation": {
        "type": "object",
        "properties": {
          "conversationId": {
            "type": "string"
          },
          "conversationName": {
            "type": "string"
          },
          "createdDate": {
            "type": "string"
          },
          "messages": {
            "nullable": true,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Message"
            }
          }
        },
        "required": [
          "conversationId",
          "conversationName",
          "createdDate",
          "messages"
        ]
      },
      "CreateConversation": {
        "type": "object",
        "properties": {
          "conversationName": {
            "type": "string"
          }
        },
        "required": [
          "conversationName"
        ]
      },
      "CreateMessage": {
        "type": "object",
        "properties": {
          "messageText": {
            "type": "string"
          }
        },
        "required": [
          "messageText"
        ]
      },
      "CreateDatasource": {
        "type": "object",
        "properties": {
          "datasourceTypeId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "datasourceTypeId",
          "name"
        ]
      },
      "Datasource": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "pathToChromaFile": {
            "type": "string"
          },
          "createdDate": {
            "type": "string"
          },
          "datasourceTypeId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "pathToChromaFile",
          "createdDate",
          "datasourceTypeId",
          "name"
        ]
      },
      "CreateDocument": {
        "type": "object",
        "properties": {
          "text": {
            "type": "string"
          }
        },
        "required": [
          "text"
        ]
      },
      "Document": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "createDate": {
            "type": "string"
          },
          "updatedDate": {
            "type": "string"
          },
          "text": {
            "type": "string"
          },
          "metadata": {
            "type": "object"
          },
          "filePath": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "createDate",
          "updatedDate",
          "text",
          "metadata",
          "filePath"
        ]
      }
    }
  }
}